freelancer-matching-platform/src/test/java/com/example/freelancer_matching_platform/controller/ProjectControllerTest.java
package com.example.freelancer_matching_platform.controller;

import com.example.freelancer_matching_platform.model.Project;
import com.example.freelancer_matching_platform.service.ProjectService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Collections;

import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(ProjectController.class)
public class ProjectControllerTest {

    @Mock
    private ProjectService projectService;

    @InjectMocks
    private ProjectController projectController;

    private MockMvc mockMvc;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
        this.mockMvc = MockMvcBuilders.standaloneSetup(projectController).build();
    }

    @Test
    void testGetAllProjects() throws Exception {
        Project project = new Project("Sample Project", "Description");
        when(projectService.getAllProjects()).thenReturn(Collections.singletonList(project));

        mockMvc.perform(get("/api/projects"))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$[0].name").value("Sample Project"));
    }

    @Test
    void testCreateProject() throws Exception {
        Project project = new Project("New Project", "New Description");
        when(projectService.saveProject(any(Project.class))).thenReturn(project);

        mockMvc.perform(post("/api/projects")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content("{\"name\": \"New Project\", \"description\": \"New Description\"}"))
                .andExpect(status().isCreated())
                .andExpect(jsonPath("$.name").value("New Project"));
    }
}






